app/requirements.txt
Flask
Flask-PyMongo
pandas
requests